requirements.txt

channels==3.0.4
channels-redis==4.0.0
daphne

dj-database-url==1.0.0

dj-static==0.0.6
Django==4.1.2

django-cors-headers==3.13.0

django-dotenv==1.4.2

django-rest-swagger==2.2.0

djangorestframework==3.14.0
drf-yasg==1.21.4
environs==9.5.0
flake8==6.0.0
gunicorn==20.1.0

Jinja2==3.1.2

psycopg2==2.9.5
psycopg2-binary==2.9.5

PyJWT==2.6.0

python-dotenv==0.21.0

redis==4.4.2

sqlparse==0.4.3

whitenoise==6.2.0

djangorestframework-camel-case

websockets==10.4
